--- a/changelog.d/2459.change.rst
+++ b/changelog.d/2459.change.rst
@@ -1,0 +1,1 @@
+Tests now run in parallel via pytest-xdist, completing in about half the time. Special thanks to :user:`webknjaz` for hard work implementing test isolation. To run without parallelization, disable the plugin with ``tox -- -p no:xdist``.