--- a/pytest.ini
+++ b/pytest.ini
@@ -4,9 +4,6 @@ addopts=
 	--doctest-modules
 	--doctest-glob=pkg_resources/api_tests.txt
 	-r sxX
-
-	# `pytest-xdist`:
-	-n auto
 doctest_optionflags=ALLOW_UNICODE ELLIPSIS
 # workaround for warning pytest-dev/pytest#6178
 junit_family=xunit2