--- a/setup.cfg
+++ b/setup.cfg
@@ -58,6 +58,7 @@ testing =
     paver
     pip>=19.1 # For proper file:// URLs support.
     jaraco.envs
+    pytest-xdist
 
 docs =
     # Keep these in sync with docs/requirements.txt