--- a/setup.cfg
+++ b/setup.cfg
@@ -47,7 +47,7 @@ testing =
 	pytest-black >= 0.3.7; python_implementation != "PyPy"
 	pytest-cov
 	pytest-mypy; python_implementation != "PyPy"
-	pytest-enabler
+	pytest-enabler >= 1.0.1
 
 	# local
     mock