--- a/setuptools/tests/test_virtualenv.py
+++ b/setuptools/tests/test_virtualenv.py
@@ -85,7 +85,7 @@ def _get_pip_versions():
 
 
 @pytest.mark.parametrize('pip_version', _get_pip_versions())
-def test_pip_upgrade_from_source(pip_version, virtualenv):
+def test_pip_upgrade_from_source(pip_version, tmp_src, virtualenv):
     """
     Check pip can upgrade setuptools from source.
     """
@@ -104,7 +104,7 @@ def test_pip_upgrade_from_source(pip_version, virtualenv):
     virtualenv.run(' && '.join((
         'python setup.py -q sdist -d {dist}',
         'python setup.py -q bdist_wheel -d {dist}',
-    )).format(dist=dist_dir), cd=SOURCE_DIR)
+    )).format(dist=dist_dir), cd=tmp_src)
     sdist = glob.glob(os.path.join(dist_dir, '*.zip'))[0]
     wheel = glob.glob(os.path.join(dist_dir, '*.whl'))[0]
     # Then update from wheel.