## Summary of changes

This is a demo showing how using pytest-xdist affects the test runs.

Resolves #2458

### Performance improvements

Comparing before/after, this patch makes the jobs run 2x faster in the CI and 4x faster locally.

|                    |        main        |      this PR      |
|--------------------|--------------------|-------------------|
| py36/ubuntu-latest | 240.99s  (0:04:00) | 137.93s (0:02:17) |
| pypy3/macos-latest | 1642.34s (0:27:22) | 732.48s (0:12:12) |

### Pull Request Checklist
- [x] Changes have tests
- [x] News fragment added in [`changelog.d/`].
  _(See [documentation][PR docs] for details)_


[`changelog.d/`]: https://github.com/pypa/setuptools/tree/master/changelog.d
[PR docs]:
https://setuptools.readthedocs.io/en/latest/development/developer-guide.html#making-a-pull-request