@webknjaz 2020-11-18
This PR demonstrates both points raised in #2458:
1) There's some race condition supposedly because of insufficient test isolation
2) It's faster with xdist. Compare `233.77s (0:03:53)` on master vs `105.42s (0:01:45)` in this PR for the GHA ubuntu py3.8 job. (https://github.com/pypa/setuptools/runs/1419466762#step:17:1076 vs https://github.com/pypa/setuptools/pull/2459/checks?check_run_id=1419764437#step:17:2199)