@webknjaz 2021-01-01
> setuptools/tests/test_build_meta.py::TestBuildMetaBackend::test_build_sdist_relative_path_import

This test seems to fail because `pytest-xdists` injects `''` into `sys.path` (it's not there when I add `-n 0` to disable it). But I've also noticed a `chdir()` call that is not thread-safe + a deprecated `tmpdir` fixture is used (it's recommended to use `tmp_path` instead).