@jaraco 2021-01-18
I've already stumbled onto a use-case this breaks. I added a `breakpoint()` to the code to inspect something. Before adding xdist, such a breakpoint would enable me to begin inspecting the code at that point. With xdist, that causes the test suite to fail with a BdbQuit exception. The tests aren't just running in parallel, but they're being forked to other processes that don't have a terminal with which to interact.