@pganssle 2021-01-18
> I've already stumbled onto a use-case this breaks. I added a `breakpoint()` to the code to inspect something.

I don't have any particular insight into this, but I've recently been working on Django and they have a test runner that defaults to running the tests in parallel and explicitly throws an error if you run `--pdb`. Might be that you simply have to disable running in parallel if you want breakpoints. Seems like a fair trade to me, TBH. I run test suites much more frequently than I set breakpoints in them, and I usually only make the "can't debug under X condition" mistake once or twice per debugging session, for a large net savings of time.