@webknjaz 2021-01-18
@jaraco 

> I'd prefer if this could use pytest-enabler, because otherwise, it becomes difficult to turn this feature off or customize the setting for a particular run. e.g.:
> [...]
> I've already stumbled onto a use-case this breaks. I added a breakpoint() to the code to inspect something. Before adding xdist, such a breakpoint would enable me to begin inspecting the code at that point. With xdist, that causes the test suite to fail with a BdbQuit exception. The tests aren't just running in parallel, but they're being forked to other processes that don't have a terminal with which to interact.

In this cause, I normally use
```console
$ tox -- -n0 -s path/to/test.py::test_something
```
`-n0` disables xdist and makes it sequential + `-s` allows to use `breakpoint()` without pytest stealing stdin/stdout streams from you.


@pganssle

> I don't have any particular insight into this, but I've recently been working on Django and they have a test runner that defaults to running the tests in parallel and explicitly throws an error if you run `--pdb`. Might be that you simply have to disable running in parallel if you want breakpoints.

I believe that it's a bit better with pure pytest: if you actually use `--pdb` and not `breakpoint()`, `xdist` will automatically disable the concurrency and turn the test run into sequential.
It'll error out, though, if you will attempt to force the concurrency:
```console
$ pytest -n2 --pdb  # fails
[...]
ERROR: --pdb is incompatible with distributing tests; try using -n0 or -nauto.
$ pytest --pdb  # works
```

> Seems like a fair trade to me, TBH. I run test suites much more frequently than I set breakpoints in them, and I usually only make the "can't debug under X condition" mistake once or twice per debugging session, for a large net savings of time.

My experience is about the same. I also occasionally do `PYTHONBREAKPOINT=pudb.remote.set_trace` and connect there via telnet in a separate terminal emulator tab.