@jaraco 2021-01-18
Good points. I'm inclined to accept the detriments in favor of the benefits. I've also confirmed that simply passing `--pdb` to pytest disables xdist, so works to bypass the behavior. Also, with the use of `pytest-enabler`, `-p no:xdist` also works.